+++ passgen/diceware/src/main/kotlin/app/passwordstore/passgen/diceware/PasswordOption.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright © 2014-2026 The Android Password Store Authors. All Rights Reserved.
+ * SPDX-License-Identifier: GPL-3.0-only
+ */
+
+package app.passwordstore.passgen.diceware
+
+public enum class PasswordOption(public val key: String) {
+  Numeral("N"),
+  Capitalise("C"),
+}